% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_followup_time.R
\name{calc_followup_time}
\alias{calc_followup_time}
\title{Mean and SD followup time for those who attended the 6 year visit.}
\usage{
calc_followup_time(data = project_data)
}
\arguments{
\item{data}{Project data}
}
\description{
Mean and SD followup time for those who attended the 6 year visit.
}
