% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renaming_fattyacids.R
\name{renaming_fa}
\alias{renaming_fa}
\alias{extract_fa_fraction}
\title{Rename fatty acid variables from PROMISE into scientific format.}
\usage{
renaming_fa(fa, keep.fraction = TRUE)

extract_fa_fraction(x, uppercase = FALSE)
}
\arguments{
\item{keep.fraction}{Logical. Whether to keep the fraction abbreviation as a
prefix.}

\item{x}{Vector (or column when through \link[dplyr:mutate]{dplyr::mutate}) of the original
fatty acid names.}

\item{uppercase}{Logical. Whether to convert the fraction abbreviation to
upper case.}
}
\value{
Reformatted fatty acid names to be scientifically accurate.
}
\description{
Fatty acids in the PROMISE cohort are in the form e.g. \code{ne160}. In the
typical data analysis pipeline, these variables may be converted into a long
format such that these variables are now rows in a column. These variables
will likely end up in plot, so it's necessary to convert them into the
appropriate scientific format (e.g. \code{NE 16:0}, or such simply \code{16:0}).
}
\section{Functions}{
\itemize{
\item \code{extract_fa_fraction}: Extract the fatty acid fraction from the PROMISE
fatty acid variables.
}}

\examples{

fattyacids <- c("ne160", "tg141n7", "TotalTG")
renaming_fa(fattyacids)
renaming_fa(fattyacids, keep.fraction = FALSE)
extract_fa_fraction(fattyacids)
extract_fa_fraction(fattyacids, uppercase = TRUE)
}
